Refactor AddClientCommandTest.java (syntax errors)
